requirements.txt
boto3==1.4.7
Pillow==4.3.0

